+++ packages/web/design-system/src/index.ts
@@ -1,10 +1,2 @@
-import enLocales from './locales/en-US.json'
-import nlLocales from './locales/nl-BE.json'
-
+export * from './locales'
 export * from './ui'
-
-// I18N LOCALES
-export const uiLocales = {
-  'en-US': enLocales,
-  'nl-BE': nlLocales,
-}