+++ packages/web/design-system/src/ui/dot/dot.style.ts
@@ -1,0 +1,22 @@
+import { tv } from '@/styles/tailwindVariants.lib'
+
+export const dotVariants = tv({
+  base: 'shrink-0 rounded-full',
+  variants: {
+    color: {
+      blue: 'bg-blue-500',
+      brand: 'bg-brand-500',
+      error: 'bg-error-500',
+      gray: 'bg-gray-400',
+      pink: 'bg-pink-500',
+      purple: 'bg-purple-500',
+      success: 'bg-success-500',
+      warning: 'bg-warning-500',
+    },
+    size: {
+      lg: 'size-2',
+      md: 'size-1.5',
+      sm: 'size-1',
+    },
+  },
+})