+++ packages/web/design-system/src/ui/dot/index.ts
@@ -1,0 +1,4 @@
+export type {
+  DotColor as UIDotColor, DotProps as UIDotProps, DotSize as UIDotSize,
+} from './dot.props'
+export { default as UIDot } from './Dot.vue'