+++ packages/web/design-system/src/ui/tabs/tabs.props.ts
@@ -4,6 +4,7 @@ import type { RouteLocationRaw } from 'vue-router'
 import type { DisabledWithReason } from '@/types/disabledWithReason.type'
 
 export type TabsVariant = 'button-border' | 'button-brand' | 'underline'
+export type TabsHorizontalListPadding = 'lg' | 'md' | 'none' | 'sm' | 'xl'
 
 export interface TabsProps {
   /**
@@ -16,6 +17,11 @@ export interface TabsProps {
    * @default 'horizontal'
    */
   orientation?: 'horizontal' | 'vertical'
+
+  /**
+   * Controls the horizontal padding of the scroll container. Only applies to the `underline` variant.
+   */
+  underlineTabsHorizontalListPadding?: TabsHorizontalListPadding
   /**
    * Defines the visual style of the tabs.
    * @default 'underline'